+++ examples/flutter_embedding/lib/components/app.dart
@@ -2,7 +2,7 @@ import 'package:jaspr/jaspr.dart';
 import 'package:jaspr_riverpod/jaspr_riverpod.dart';
 
 import 'effects_controls.dart';
-import 'flutter_app_container.dart' if (dart.library.html) 'flutter_app_container_web.dart';
+import 'flutter_app_container.dart' if (dart.library.js_interop) 'flutter_app_container_web.dart';
 import 'interop_controls.dart';
 
 @client