+++ packages/jaspr/lib/src/browser/clients.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
-import 'dart:html';
+
+import 'package:web/web.dart' as web;
 
 import '../foundation/marker_utils.dart';
 import '../framework/framework.dart';
@@ -16,7 +17,7 @@ void registerClientsSync(Map<String, ClientBuilder> clients) {
 void registerClients(Map<String, ClientLoader> clients) {
   final Map<String, ClientBuilder> builders = {};
   _applyClients((name) {
-    var client = (builders[name] ?? clients[name]!()) as FutureOr<ClientBuilder>;
+    final client = (builders[name] ?? clients[name]!()) as FutureOr<ClientBuilder>;
     if (client is ClientBuilder) {
       return builders[name] = client;
     } else {
@@ -37,36 +38,38 @@ final _compStartRegex = RegExp('^$clientMarkerPrefixRegex(\\S+)(?:\\s+data=(.*))
 final _compEndRegex = RegExp('^/$clientMarkerPrefixRegex(\\S+)\$');
 
 void _applyClients(FutureOr<ClientBuilder> Function(String) fn) {
-  var iterator = NodeIterator(document, NodeFilter.SHOW_COMMENT);
+  var iterator = web.document.createNodeIterator(web.document, 128 /* NodeFilter.SHOW_COMMENT */);
 
-  List<(String, String?, Node)> nodes = [];
+  List<(String, String?, web.Node)> nodes = [];
 
-  Comment? currNode;
-  while ((currNode = iterator.nextNode() as Comment?) != null) {
-    var value = currNode!.nodeValue ?? '';
+  web.Comment? currNode;
+  while ((currNode = iterator.nextNode() as web.Comment?) != null) {
+    final value = currNode!.nodeValue ?? '';
     var match = _compStartRegex.firstMatch(value);
     if (match != null) {
-      var name = match.group(1)!;
-      var data = match.group(2);
+      final name = match.group(1)!;
+      final data = match.group(2);
 
       nodes.add((name, data, currNode));
     }
 
     match = _compEndRegex.firstMatch(value);
     if (match != null) {
-      var name = match.group(1)!;
+      final name = match.group(1)!;
 
       if (nodes.last.$1 == name) {
-        var comp = nodes.removeLast();
-        var start = comp.$3;
+        final comp = nodes.removeLast();
+        final start = comp.$3;
         assert(start.parentNode == currNode.parentNode);
 
-        var between = (start, currNode);
+        final between = (start, currNode);
 
         // Remove the data string.
         start.text = '$clientMarkerPrefix${comp.$1}';
 
-        Map<String, dynamic> params = comp.$2 != null ? jsonDecode(unescapeMarkerText(comp.$2!)) : {};
+        final params = comp.$2 != null //
+            ? jsonDecode(unescapeMarkerText(comp.$2!)) as Map<String, dynamic>
+            : <String, dynamic>{};
         unawaited(_runBuilder(name, fn(name), params, between));
       }
     }
@@ -77,7 +80,7 @@ Future<void> _runBuilder(
   String name,
   FutureOr<ClientBuilder> builder,
   Map<String, dynamic> params,
-  (Node, Node) between,
+  (web.Node, web.Node) between,
 ) async {
   if (builder is Future<ClientBuilder>) {
     builder = await builder;