+++ packages/jaspr/test/browser/head/head_browser_test.dart
@@ -1,7 +1,5 @@
 @TestOn('browser')
 
-import 'dart:html';
-
 import 'package:jaspr/src/components/document/document_client.dart';
 import 'package:jaspr_test/browser_test.dart';
 
@@ -14,18 +12,21 @@ void main() {
 
       var nodes = AttachAdapter.instanceFor(AttachTarget.head).liveNodes.toList();
 
-      expect(nodes, hasLength(3));
-      expect((nodes[0] as Element).outerHtml, equals('<title>c</title>'));
-      expect((nodes[1] as Element).outerHtml, equals('<meta name="test" content="b">'));
-      expect((nodes[2] as Element).outerHtml, equals('<meta name="c" content="e">'));
+      expect(nodes, [
+        hasOuterHtml('<title>c</title>'),
+        hasOuterHtml('<meta name="test" content="b">'),
+        hasOuterHtml('<meta name="c" content="e">'),
+      ]);
 
       await tester.click(find.tag('button'));
 
       nodes = AttachAdapter.instanceFor(AttachTarget.head).liveNodes.toList();
 
-      expect((nodes[0] as Element).outerHtml, equals('<title>d</title>'));
-      expect((nodes[1] as Element).outerHtml, equals('<meta name="test" content="b">'));
-      expect((nodes[2] as Element).outerHtml, equals('<meta name="c" content="d">'));
+      expect(nodes, [
+        hasOuterHtml('<title>d</title>'),
+        hasOuterHtml('<meta name="test" content="b">'),
+        hasOuterHtml('<meta name="c" content="d">'),
+      ]);
     });
   });
 }