+++ packages/jaspr/CHANGELOG.md
@@ -1,4 +1,6 @@
-## Unreleased patch
+## Unreleased breaking
+
+- **BREAKING** Migrate all packages to `package:web`, replacing `dart:html`.
 
 - Add `InheritedModel<T>` similar to Flutters [InheritedModel](https://api.flutter.dev/flutter/widgets/InheritedModel-class.html)
 - Fixed server issue during tests where the web directory would never resolve.